Implement Google Maps geolocation for listings and auto-link to nearest university
- Integrated Google Maps API wrapper to fetch coordinates from listing address
- Updated AccommodationService to calculate nearest university based on coordinates
- Modified DTOs and Edit/Create Razor pages to support location-aware logic
- Refactored repositories and services to align with new geolocation flow
- Added logging and error handling for address validation

+++ BLL/DTOs/Accommodation/AccommodationBookingDto.cs
@@ -19,7 +19,7 @@ namespace BLL.DTOs.Accommodation
         public decimal MonthlyRent { get; set; }
         public bool IsAvailable { get; set; }
         public int MaxOccupants { get; set; }
-        public int Size { get; set; }
+        public decimal Size { get; set; }
         public DateTime AvailableFrom { get; set; }
         public string AccommodationType { get; set; } = string.Empty;
         public string UniversityName { get; set; } = string.Empty;
